requirements.txt
pandas==2.2.3
numpy==2.1.3
statsmodels==0.14.4
scikit-learn==1.5.2
matplotlib==3.10.0
seaborn==0.13.2
flask==3.1.0
openpyxl==3.1.5
python-dotenv==1.0.1
